feat: 요구 사항에 따른 route 변경
