--- src/pages/profile/[id].tsx
@@ -1,5 +1,0 @@
-import ProfilePageLayout from "@/components/profile/ProfilePageLayout";
-
-export default function UserProfilePage() {
-	return <ProfilePageLayout />;
-}
