--- src/pages/profile/my.tsx
@@ -1,5 +1,0 @@
-import ProfilePageLayout from "@/components/profile/ProfilePageLayout";
-
-export default function MyProfilePage() {
-	return <ProfilePageLayout />;
-}
